+++ client/src/api/block_builder/input_selection/helpers.rs
@@ -5,89 +5,34 @@
 
 use iota_types::block::{
     address::{Address, AliasAddress, Ed25519Address, NftAddress},
-    output::{
-        unlock_condition::{AddressUnlockCondition, StorageDepositReturnUnlockCondition},
-        BasicOutputBuilder, NativeTokens, NativeTokensBuilder, Output, Rent, RentStructure, UnlockCondition,
-    },
+    output::Output,
 };
 
-use crate::{
-    api::input_selection::{get_minted_and_melted_native_tokens, types::AccumulatedOutputAmounts},
-    secret::types::InputSigningData,
-    Result,
-};
+use crate::secret::types::InputSigningData;
 
-// Calculate required accumulated amounts from the outputs, considers also minted and melted native tokens
-pub(crate) fn get_accumulated_output_amounts<'a>(
-    inputs: &(impl Iterator<Item = &'a Output> + Clone),
-    outputs: impl Iterator<Item = &'a Output> + Clone,
-) -> Result<AccumulatedOutputAmounts> {
-    // Calculate the total tokens to spend
-    let mut required_amount: u64 = 0;
-    let mut required_native_tokens = NativeTokensBuilder::new();
+// Dedup inputs by output id, because other data could be different, even if it's the same output
+// TODO remove ?
+// pub(crate) fn dedup_inputs(
+//     mandatory_inputs: &mut Vec<InputSigningData>,
+//     additional_inputs: &mut Vec<InputSigningData>,
+// ) {
+//     // Sorting inputs by OutputId so duplicates can be safely removed.
+//     mandatory_inputs.sort_by_key(|input| *input.output_metadata.output_id());
+//     mandatory_inputs.dedup_by_key(|input| *input.output_metadata.output_id());
+//     additional_inputs.sort_by_key(|input| *input.output_metadata.output_id());
+//     additional_inputs.dedup_by_key(|input| *input.output_metadata.output_id());
 
-    for output in outputs.clone() {
-        required_amount += output.amount();
-
-        if let Some(output_native_tokens) = output.native_tokens() {
-            required_native_tokens.add_native_tokens(output_native_tokens.clone())?;
-        }
-    }
-
-    // check if a foundry mints or melts native tokens
-    let (minted_native_tokens, melted_native_tokens) = get_minted_and_melted_native_tokens(inputs, outputs)?;
-    // add melted native tokens as outputs, because we need to have this amount in the inputs
-    required_native_tokens.merge(melted_native_tokens)?;
-
-    Ok(AccumulatedOutputAmounts {
-        minted_native_tokens,
-        amount: required_amount,
-        native_tokens: required_native_tokens,
-    })
-}
-
-/// Computes the minimum storage deposit amount that a basic output needs to have with an [AddressUnlockCondition] and
-/// optional [NativeTokens].
-pub fn minimum_storage_deposit_basic_output(
-    config: &RentStructure,
-    address: &Address,
-    native_tokens: &Option<NativeTokens>,
-    token_supply: u64,
-) -> Result<u64> {
-    let address_condition = UnlockCondition::Address(AddressUnlockCondition::new(*address));
-    let mut basic_output_builder = BasicOutputBuilder::new_with_amount(Output::AMOUNT_MIN)?;
-    if let Some(native_tokens) = native_tokens {
-        basic_output_builder = basic_output_builder.with_native_tokens(native_tokens.clone());
-    }
-    let basic_output = basic_output_builder
-        .add_unlock_condition(address_condition)
-        .finish_output(token_supply)?;
-
-    Ok(basic_output.rent_cost(config))
-}
-
-/// Get the `StorageDepositReturnUnlockCondition`, if not expired
-pub(crate) fn sdr_not_expired(output: &Output, current_time: u32) -> Option<&StorageDepositReturnUnlockCondition> {
-    if let Some(unlock_conditions) = output.unlock_conditions() {
-        if let Some(sdr) = unlock_conditions.storage_deposit_return() {
-            let expired = if let Some(expiration) = unlock_conditions.expiration() {
-                current_time >= expiration.timestamp()
-            } else {
-                false
-            };
-
-            // We only have to send the storage deposit return back if the output is not expired
-            if !expired { Some(sdr) } else { None }
-        } else {
-            None
-        }
-    } else {
-        None
-    }
-}
+//     // Remove additional inputs that are already mandatory.
+//     // TODO: could be done more efficiently with itertools unique?
+//     additional_inputs.retain(|input| {
+//         !mandatory_inputs
+//             .iter()
+//             .any(|mandatory_input| input.output_metadata.output_id() == mandatory_input.output_metadata.output_id())
+//     });
+// }
 
 // Inputs need to be sorted before signing, because the reference unlock conditions can only reference a lower index
-pub(crate) fn sort_input_signing_data(inputs: Vec<InputSigningData>) -> crate::Result<Vec<InputSigningData>> {
+pub(crate) fn _sort_input_signing_data(inputs: Vec<InputSigningData>) -> crate::Result<Vec<InputSigningData>> {
     // filter for ed25519 address first, safe to unwrap since we encoded it before
     let (mut sorted_inputs, alias_nft_address_inputs): (Vec<InputSigningData>, Vec<InputSigningData>) = inputs
         .into_iter()