+++ client/CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Complete refactoring of Input Selection;
 - Updated dependencies;
 - Fields of `Error::InputAddressNotFound` are now named;
 - `Event` renamed to `ParticipationEvent`;